requirements.txt
beautifulsoup4==4.12.2
pytest==7.4.3
python-dotenv==1.0.0
selenium==4.15.1
